requirements.txt
datasets==2.16.1
openai==1.30.1
PyPDF2==3.0.1
transformers==4.37.2
langchain_experimental
langchain_openai
python-dotenv==1.0.1
pyyaml==6.0.1
coloredlogs==15.0.1
mdc==1.2.1
pytest==8.1.2
wand
PyCryptodome
azure-identity
torch
python-pptx
promptflow.core
promptflow.eval
jsonlines